Move some support files to support/
