+++ docker-compose.yml
@@ -4,7 +4,7 @@ services:
   web:
     build:
       context: .
-      dockerfile: Dockerfile
+      dockerfile: support/Dockerfile
     volumes:
       - .:/app
       - node_modules:/app/node_modules
@@ -16,7 +16,7 @@ services:
     environment:
       CHOKIDAR_USEPOLLING: 'true'
       PORT: '19336'
-    command: "sh ./docker-compose.web.sh"
+    command: 'sh ./support/start.sh'
 
 volumes:
   node_modules:
