+++ gatsby-node.js
@@ -1,3 +1,3 @@
-require('./register')()
+require('./support/register')()
 
 module.exports = require('./gatsby-node.ts')
