+++ support/register.js
